+++ apis/appstream/v1beta1/zz_fleet_types.go
@@ -16,29 +16,29 @@ import (
 type ComputeCapacityInitParameters struct {
 
 	// Desired number of streaming instances.
-	DesiredInstances *float64 `json:"desiredInstances,omitempty" tf:"desired_instances,omitempty"`
+	DesiredInstances *int64 `json:"desiredInstances,omitempty" tf:"desired_instances,omitempty"`
 }
 
 type ComputeCapacityObservation struct {
 
 	// Number of currently available instances that can be used to stream sessions.
-	Available *float64 `json:"available,omitempty" tf:"available,omitempty"`
+	Available *int64 `json:"available,omitempty" tf:"available,omitempty"`
 
 	// Desired number of streaming instances.
-	DesiredInstances *float64 `json:"desiredInstances,omitempty" tf:"desired_instances,omitempty"`
+	DesiredInstances *int64 `json:"desiredInstances,omitempty" tf:"desired_instances,omitempty"`
 
 	// Number of instances in use for streaming.
-	InUse *float64 `json:"inUse,omitempty" tf:"in_use,omitempty"`
+	InUse *int64 `json:"inUse,omitempty" tf:"in_use,omitempty"`
 
 	// Total number of simultaneous streaming instances that are running.
-	Running *float64 `json:"running,omitempty" tf:"running,omitempty"`
+	Running *int64 `json:"running,omitempty" tf:"running,omitempty"`
 }
 
 type ComputeCapacityParameters struct {
 
 	// Desired number of streaming instances.
 	// +kubebuilder:validation:Optional
-	DesiredInstances *float64 `json:"desiredInstances" tf:"desired_instances,omitempty"`
+	DesiredInstances *int64 `json:"desiredInstances" tf:"desired_instances,omitempty"`
 }
 
 type DomainJoinInfoInitParameters struct {
@@ -79,7 +79,7 @@ type FleetInitParameters struct {
 	Description *string `json:"description,omitempty" tf:"description,omitempty"`
 
 	// Amount of time that a streaming session remains active after users disconnect.
-	DisconnectTimeoutInSeconds *float64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
+	DisconnectTimeoutInSeconds *int64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
 
 	// Human-readable friendly name for the AppStream fleet.
 	DisplayName *string `json:"displayName,omitempty" tf:"display_name,omitempty"`
@@ -94,7 +94,7 @@ type FleetInitParameters struct {
 	FleetType *string `json:"fleetType,omitempty" tf:"fleet_type,omitempty"`
 
 	// Amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the disconnect_timeout_in_seconds time interval begins.
-	IdleDisconnectTimeoutInSeconds *float64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
+	IdleDisconnectTimeoutInSeconds *int64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
 
 	// ARN of the public, private, or shared image to use.
 	ImageArn *string `json:"imageArn,omitempty" tf:"image_arn,omitempty"`
@@ -106,7 +106,7 @@ type FleetInitParameters struct {
 	InstanceType *string `json:"instanceType,omitempty" tf:"instance_type,omitempty"`
 
 	// Maximum amount of time that a streaming session can remain active, in seconds.
-	MaxUserDurationInSeconds *float64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
+	MaxUserDurationInSeconds *int64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
 
 	// Unique name for the fleet.
 	Name *string `json:"name,omitempty" tf:"name,omitempty"`
@@ -136,7 +136,7 @@ type FleetObservation struct {
 	Description *string `json:"description,omitempty" tf:"description,omitempty"`
 
 	// Amount of time that a streaming session remains active after users disconnect.
-	DisconnectTimeoutInSeconds *float64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
+	DisconnectTimeoutInSeconds *int64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
 
 	// Human-readable friendly name for the AppStream fleet.
 	DisplayName *string `json:"displayName,omitempty" tf:"display_name,omitempty"`
@@ -157,7 +157,7 @@ type FleetObservation struct {
 	ID *string `json:"id,omitempty" tf:"id,omitempty"`
 
 	// Amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the disconnect_timeout_in_seconds time interval begins.
-	IdleDisconnectTimeoutInSeconds *float64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
+	IdleDisconnectTimeoutInSeconds *int64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
 
 	// ARN of the public, private, or shared image to use.
 	ImageArn *string `json:"imageArn,omitempty" tf:"image_arn,omitempty"`
@@ -169,7 +169,7 @@ type FleetObservation struct {
 	InstanceType *string `json:"instanceType,omitempty" tf:"instance_type,omitempty"`
 
 	// Maximum amount of time that a streaming session can remain active, in seconds.
-	MaxUserDurationInSeconds *float64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
+	MaxUserDurationInSeconds *int64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
 
 	// Unique name for the fleet.
 	Name *string `json:"name,omitempty" tf:"name,omitempty"`
@@ -201,7 +201,7 @@ type FleetParameters struct {
 
 	// Amount of time that a streaming session remains active after users disconnect.
 	// +kubebuilder:validation:Optional
-	DisconnectTimeoutInSeconds *float64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
+	DisconnectTimeoutInSeconds *int64 `json:"disconnectTimeoutInSeconds,omitempty" tf:"disconnect_timeout_in_seconds,omitempty"`
 
 	// Human-readable friendly name for the AppStream fleet.
 	// +kubebuilder:validation:Optional
@@ -235,7 +235,7 @@ type FleetParameters struct {
 
 	// Amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the disconnect_timeout_in_seconds time interval begins.
 	// +kubebuilder:validation:Optional
-	IdleDisconnectTimeoutInSeconds *float64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
+	IdleDisconnectTimeoutInSeconds *int64 `json:"idleDisconnectTimeoutInSeconds,omitempty" tf:"idle_disconnect_timeout_in_seconds,omitempty"`
 
 	// ARN of the public, private, or shared image to use.
 	// +kubebuilder:validation:Optional
@@ -251,7 +251,7 @@ type FleetParameters struct {
 
 	// Maximum amount of time that a streaming session can remain active, in seconds.
 	// +kubebuilder:validation:Optional
-	MaxUserDurationInSeconds *float64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
+	MaxUserDurationInSeconds *int64 `json:"maxUserDurationInSeconds,omitempty" tf:"max_user_duration_in_seconds,omitempty"`
 
 	// Unique name for the fleet.
 	// +kubebuilder:validation:Optional
